requirements.txt
addfips>=0.4.0
dask>=2.8
dask[array]>=2.8
fiona>=1.9.0
geopandas>=0.8
NREL-gaps>=0.3.3
NREL-reV>=0.8.4
NREL-rex>=0.2.80
psycopg2-binary>=2.8
pydantic>=2.5.3
pyproj>=3.0.1
pywavelets>=1.0
rasterio>=1.2.10
scikit-image>=0.18
scikit-learn>=0.22
shapely<2.0.0
toolz>=0.7.3
